feat!: add duckdb storage backend
