+++ docs/source/api/api/storage/anyvar.storage.duckdb.rst
@@ -1,0 +1,9 @@
+﻿anyvar.storage.duckdb
+=====================
+
+.. automodule:: anyvar.storage.duckdb
+   :members:
+   :no-inherited-members:
+   :undoc-members:
+   :special-members: __init__
+   :exclude-members: model_fields, model_config, metadata
