+++ docs/source/api/api/storage/anyvar.storage.snowflake.rst
@@ -1,0 +1,9 @@
+﻿anyvar.storage.snowflake
+========================
+
+.. automodule:: anyvar.storage.snowflake
+   :members:
+   :no-inherited-members:
+   :undoc-members:
+   :special-members: __init__
+   :exclude-members: model_fields, model_config, metadata
