+++ docs/source/configuration/storage.rst
@@ -4,7 +4,7 @@ Object Storage Configuration
 Storage Connection
 ==================
 
-Use the ``ANYVAR_STORAGE_URI`` environment variable to pass a `libpq connection string <https://www.postgresql.org/docs/current/libpq.html>`_ to the PostgreSQL connection constructor, or an empty string (``""``) to use :ref:`stateless mode <stateless_mode>`.
+AnyVar supports several different database engines. Use the ``ANYVAR_STORAGE_URI`` environment variable to designate the database engine and its location.
 
 .. list-table::
    :widths: 30 70
@@ -14,6 +14,13 @@ Use the ``ANYVAR_STORAGE_URI`` environment variable to pass a `libpq connection 
      - Default Value
    * - ``ANYVAR_STORAGE_URI``
      - ``"postgresql://postgres@localhost:5432/anyvar"``
+
+
+* A `libpq connection string <https://www.postgresql.org/docs/current/libpq.html>`_ constructs a PostgreSQL connection. This is the default database engine and is recommended for most use cases.
+* A URI with a ``snowflake://`` scheme constructs a Snowflake connection. See the :ref:`AnyVar Snowflake configuration <snowflake>` page for more information.
+* a URI with a ``duckdb://`` scheme constructs a `DuckDB <https://duckdb.org/>`_ connection. Use a relative file path, eg ``duckdb:///my_variants.duckdb``, for a file-based instance, or ``duckdb:///:memory:`` for an in-memory instance. See the :py:class:`~anyvar.storage.duckdb` API reference page for more information.
+* An empty string (ie ``export ANYVAR_STORAGE_URI=""``) enables :ref:`stateless mode <stateless_mode>`.
+
 
 
 Rename Tables
