+++ src/anyvar/anyvar.py
@@ -36,6 +36,7 @@ def create_storage(uri: str | None = None) -> Storage:
 
     * PostgreSQL: ``postgresql://[username]:[password]@[domain]/[database]``
     * Snowflake: ``snowflake://sf_username:@sf_account_identifier/sf_db_name/sf_schema_name?password=sf_password``
+    * DuckDB: ``duckdb:///:memory:` or ``duckdb:///relative/path/to/file``
 
     For no database (for testing or non-persistent use cases), use an empty string.
 
@@ -52,6 +53,10 @@ def create_storage(uri: str | None = None) -> Storage:
         from anyvar.storage.snowflake import SnowflakeObjectStore  # noqa: PLC0415
 
         storage = SnowflakeObjectStore(uri)
+    elif parsed_uri.scheme == "duckdb":
+        from anyvar.storage.duckdb import DuckDbObjectStore  # noqa: PLC0415
+
+        storage = DuckDbObjectStore(uri)
     elif parsed_uri.scheme == "":
         from anyvar.storage.no_db import NoObjectStore  # noqa: PLC0415
