+++ src/anyvar/storage/postgres.py
@@ -1,47 +1,25 @@
 """Provide PostgreSQL-based storage implementation."""
 
 import json
-import logging
-from collections import defaultdict
-from collections.abc import Iterable
 
-from ga4gh.vrs import models as vrs_models
 from pydantic import JsonValue
-from sqlalchemy import and_, create_engine, delete, func, or_, select
-from sqlalchemy.dialects.postgresql import insert
-from sqlalchemy.exc import IntegrityError
-from sqlalchemy.orm import joinedload, sessionmaker
-
-from anyvar.core import metadata
-from anyvar.core import objects as anyvar_objects
-from anyvar.storage import orm
-from anyvar.storage.base import (
-    AlleleSearchPage,
-    DataIntegrityError,
-    IncompleteVrsObjectError,
-    InvalidSearchParamsError,
-    Storage,
+from sqlalchemy import (
+    ColumnElement,
+    Engine,
+    Index,
+    create_engine,
+    delete,
+    func,
 )
-from anyvar.storage.mapper_registry import mapper_registry
-from anyvar.storage.orm import create_tables
+from sqlalchemy.dialects.postgresql import insert
+from sqlalchemy.orm import Session, sessionmaker
 
-_logger = logging.getLogger(__name__)
+from anyvar.storage import orm
+from anyvar.storage.sqlalchemy import SqlAlchemyStorage
 
 
-class PostgresObjectStore(Storage):
-    """PostgreSQL storage backend using dedicated ORM tables.
-
-    This implementation uses the new Allele, Location, and SequenceReference tables
-    with object mapping to convert between VRS models and database entities.
-    """
-
-    MAX_ROWS = 100
-
-    _VRS_OBJECT_INSERT_ORDER: list[str] = [  # noqa: RUF012
-        orm.SequenceReference.__name__,
-        orm.Location.__name__,
-        orm.Allele.__name__,
-    ]
+class PostgresObjectStore(SqlAlchemyStorage):
+    """PostgreSQL storage backend using dedicated ORM tables."""
 
     def __init__(self, db_url: str, *args, **kwargs) -> None:
         """Initialize PostgreSQL storage.
@@ -52,7 +30,42 @@ class PostgresObjectStore(Storage):
         self.engine = create_engine(db_url)
         self.session_factory = sessionmaker(bind=self.engine)
         self.batch_size = kwargs.get("batch_size", 1000)
-        create_tables(self.db_url)
+        self._initialize(self.engine)
+
+    def _initialize(self, engine: Engine) -> None:
+        """Initialize postgres
+
+        Ensure existence of tables, engine-specific indices, etc
+        """
+        orm.Base.metadata.create_all(bind=engine)
+        self._create_indices(engine)
+
+    def _create_indices(self, engine: Engine) -> None:
+        """Create postgres-specific indices"""
+        indices = [
+            Index(
+                "ix_location_ref_overlap",
+                orm.Location.sequence_reference_id,
+                func.int8range(
+                    orm.Location.start,
+                    orm.Location.end,
+                    "[]",
+                ),
+                postgresql_using="gist",
+            ),
+            Index(
+                "idx_extensions_object_id_name",
+                orm.Extension.object_id,
+                orm.Extension.name,
+            ),
+            Index(
+                "idx_mappings_dest_id",
+                orm.VariationMapping.dest_id,
+            ),
+        ]
+
+        for idx in indices:
+            idx.create(bind=engine, checkfirst=True)
 
     def close(self) -> None:
         """Close the storage backend."""
@@ -62,278 +75,31 @@ class PostgresObjectStore(Storage):
         NOTE: This is a no-op for synchronous storage backends.
         """
 
-    def wipe_db(self) -> None:
-        """Wipe all data from the storage backend."""
-        with self.session_factory() as session, session.begin():
-            # Delete all data from tables in dependency order
-            session.execute(delete(orm.VariationMapping))
-            session.execute(delete(orm.Allele))
-            session.execute(delete(orm.Location))
-            session.execute(delete(orm.SequenceReference))
-
-            # Delete other tables
-            session.execute(delete(orm.VrsObject))
-            session.execute(delete(orm.Extension))
-
-    def add_objects(self, objects: Iterable[anyvar_objects.SupportedVrsObject]) -> None:
-        """Add multiple VRS objects to storage using bulk inserts.
-
-        If an object ID conflicts with an existing object, skip it.
-
-        This method assumes that for VRS objects (e.g. `Allele`, `SequenceLocation`,
-        `SequenceReference`) the `.id` property is present and uses the correct
-        GA4GH identifier for that object. It also assumes that contained objects are
-        similarly properly identified and materialized in full, not just as an IRI reference.
-        An error is raised if these assumptions are violated, rolling back the entire
-        transaction.
-
-        :param objects: VRS objects to add to storage
-        :raise IncompleteVrsObjectError: if object is missing required properties or if
-            required properties aren't fully dereferenced
-        """
-        objects_list: list[anyvar_objects.SupportedVrsObject] = list(objects)
-        if not objects_list:
-            return
-
-        # Collect unique entities by ID to avoid duplicates
-        vrs_objects = defaultdict(dict[str, orm.Base])
-
-        # Process all objects and extract their components
-        for vrs_object in objects_list:
-            try:
-                db_entity = mapper_registry.to_db_entity(vrs_object)
-            except AttributeError as e:
-                raise IncompleteVrsObjectError from e
-
-            object_parts = db_entity.disassemble()
-            for entity_type, entity in object_parts.items():
-                vrs_objects[entity_type][entity.id] = entity  # type: ignore (all children of orm.Base have an `id`)
-
-        with self.session_factory() as session, session.begin():
-            # Use ON CONFLICT DO NOTHING to handle duplicates gracefully
-            # We should have already de-duplicated by ID above, but duplicates
-            # may already exist in the database.
-
-            for vrs_object_type in self._VRS_OBJECT_INSERT_ORDER:
-                objects_by_id = vrs_objects[vrs_object_type]
-                if objects_by_id:
-                    dicts = [entity.to_dict() for entity in objects_by_id.values()]
-                    orm_model = getattr(orm, vrs_object_type)
-                    stmt = insert(orm_model).on_conflict_do_nothing()
-                    session.execute(stmt, dicts)
-
-    def get_objects(
-        self,
-        object_type: type[anyvar_objects.SupportedVrsObject],
-        object_ids: Iterable[str],
-    ) -> Iterable[anyvar_objects.SupportedVrsObject]:
-        """Retrieve multiple VRS objects from storage by their IDs.
-
-        If no object matches a given ID, that ID is skipped
-
-        :param object_type: type of object to get
-        :param object_ids: IDs of objects to fetch
-        :return: iterable collection of VRS objects matching given IDs
-        """
-        object_ids_list = list(object_ids)
-        if not object_ids_list:
-            return []
-
-        results = []
-        with self.session_factory() as session:
-            if object_type is vrs_models.Allele:
-                # Get alleles with eager loading
-                stmt = (
-                    select(orm.Allele)
-                    .options(
-                        joinedload(orm.Allele.location).joinedload(
-                            orm.Location.sequence_reference
-                        )
-                    )
-                    .where(orm.Allele.id.in_(object_ids_list))
-                )
-                db_objects = session.scalars(stmt).all()
-            elif object_type is vrs_models.SequenceLocation:
-                # Get locations with eager loading
-                stmt = (
-                    select(orm.Location)
-                    .options(joinedload(orm.Location.sequence_reference))
-                    .where(orm.Location.id.in_(object_ids_list))
-                )
-                db_objects = session.scalars(stmt).all()
-            elif object_type is vrs_models.SequenceReference:
-                # Get sequence references
-                stmt = select(orm.SequenceReference).where(
-                    orm.SequenceReference.id.in_(object_ids_list)
-                )
-
-                db_objects = session.scalars(stmt).all()
-            else:
-                raise ValueError(f"Unsupported object type: {object_type}")
-
-            for db_object in db_objects:
-                vrs_object = mapper_registry.from_db_entity(db_object)
-                results.append(vrs_object)
-
-        return results
-
-    def delete_objects(
-        self,
-        object_type: type[anyvar_objects.SupportedVrsObject],
-        object_ids: Iterable[str],
+    def _insert_ignore_conflict(
+        self, session: Session, orm_model: type[orm.Base], values: list[dict]
     ) -> None:
-        """Delete all objects of a specific type from storage.
+        """Perform an insert of the given values, ignoring ID conflicts
 
-        * If no object matching a given ID is found, it's ignored.
-        * Deletes do not cascade.
-
-        :param object_type: type of objects to delete
-        :param object_ids: IDs of objects to delete
-        :raise DataIntegrityError: if attempting to delete an object which is
-            depended upon by another object
+        Should be implemented using specific engines/dialects of subclasses
         """
-        object_ids_list = list(object_ids)
+        stmt = insert(orm_model).on_conflict_do_nothing()
+        session.execute(stmt, values)
 
-        with self.session_factory() as session, session.begin():
-            if object_type is vrs_models.Allele:
-                stmt = delete(orm.Allele).where(orm.Allele.id.in_(object_ids_list))
-            elif object_type is vrs_models.SequenceLocation:
-                stmt = delete(orm.Location).where(orm.Location.id.in_(object_ids_list))
-            elif object_type is vrs_models.SequenceReference:
-                stmt = delete(orm.SequenceReference).where(
-                    orm.SequenceReference.id.in_(object_ids_list)
-                )
-            else:
-                raise ValueError(f"Unsupported object type: {object_type}")
-            try:
-                session.execute(stmt)
-            except IntegrityError as e:
-                _logger.exception(
-                    "Attempted deletion that violated a foreign key constraint"
-                )
-                raise DataIntegrityError from e
-
-    def add_mapping(self, mapping: metadata.VariationMapping) -> None:
-        """Add a mapping between two objects.
-
-        If the mapping instance already exists, do nothing.
-
-        Todo:
-        * Implement insert constraint/MissingVariationReferenceError in #286
-
-        :param mapping: mapping object
-        :raises ValueError: If source_id equals dest_id
-        :raise MissingVariationReferenceError: if source or destination IDs aren't present in DB
-
-        """
-        if mapping.source_id == mapping.dest_id:
-            msg = f"source_id cannot equal dest_id: {mapping.source_id}"
-            raise ValueError(msg)
-
-        stmt = (
-            insert(orm.VariationMapping)
-            .values(
-                [
-                    {
-                        "source_id": mapping.source_id,
-                        "dest_id": mapping.dest_id,
-                        "mapping_type": mapping.mapping_type,
-                    }
-                ]
-            )
-            .on_conflict_do_nothing()
-        )
-        try:
-            with self.session_factory() as session, session.begin():
-                session.execute(stmt)
-        except IntegrityError as e:
-            raise KeyError from e
-
-    def delete_mapping(self, mapping: metadata.VariationMapping) -> None:
-        """Delete a mapping between two objects.
-
-        * If no such mapping exists in the DB, does nothing.
-        * Deletes do not cascade.
-
-        :param mapping: mapping object
-        """
-        stmt = (
-            delete(orm.VariationMapping)
-            .where(orm.VariationMapping.source_id == mapping.source_id)
-            .where(orm.VariationMapping.dest_id == mapping.dest_id)
-            .where(orm.VariationMapping.mapping_type == mapping.mapping_type)
-        )
-        with self.session_factory() as session, session.begin():
-            session.execute(stmt)
-
-    def get_mappings(
+    def _overlaps_interval_predicate(
         self,
-        object_id: str,
-        as_source: bool,
-        mapping_type: metadata.VariationMappingType | None = None,
-    ) -> Iterable[metadata.VariationMapping]:
-        """Return an iterable of mappings
+        start: int,
+        stop: int,
+    ) -> ColumnElement[bool]:
+        """Return a PostgreSQL range-overlap predicate.
 
-        Optionally provide a type to filter results.
-
-        :param object_id: ID of object to get mappings for
-        :param as_source: If ``True``, object_id is treated as the source. If ``False``,
-            ``object_id`` is treated as the destination.
-        :param mapping_type: The type of mapping to retrieve (defaults to `None` to
-            retrieve all mappings for the source ID)
-        :return: iterable collection of mapping descriptors (empty if no matching mappings exist)
+        Uses PostgreSQL's native range types and overlap operator to enable
+        efficient interval searches via a GiST index.
         """
-        stmt = select(orm.VariationMapping).limit(self.MAX_ROWS)
-        if as_source:
-            stmt = stmt.where(orm.VariationMapping.source_id == object_id)
-        else:
-            stmt = stmt.where(orm.VariationMapping.dest_id == object_id)
-        if mapping_type:
-            stmt = stmt.where(orm.VariationMapping.mapping_type == mapping_type)
-        with self.session_factory() as session, session.begin():
-            mappings = session.scalars(stmt).all()
-            return [mapper_registry.from_db_entity(mapping) for mapping in mappings]
-
-    def add_extension(self, extension: metadata.Extension) -> None:
-        """Add an extension to the database.
-
-        Adding the same extension repeatedly creates redundant records.
-
-        Todo:
-        * Implement insert constraint/MissingVariationReferenceError in #286
-
-        :param extension: The extension to add
-        :raise MissingVariationReferenceError: if no object corresponding to the extension's object ID is present in DB
-
-        """
-        db_entity: orm.Extension = mapper_registry.to_db_entity(extension)
-        stmt = insert(orm.Extension).returning(orm.Extension.id)
-        with self.session_factory() as session, session.begin():
-            session.execute(stmt, db_entity.to_dict()).scalar_one()
-
-    def get_extensions(
-        self, object_id: str, extension_name: str | None = None
-    ) -> list[metadata.Extension]:
-        """Get all extensions for the specified object, optionally filtered by type.
-
-        :param object_id: The ID of the object to retrieve extensions for
-        :param extension_type: The type of extension to retrieve (defaults to `None` to retrieve all extensions for the object)
-        :return: A list of extensions
-        """
-        stmt = select(orm.Extension).where(orm.Extension.object_id == object_id)
-
-        if extension_name:
-            stmt = stmt.where(orm.Extension.name == extension_name)
-
-        stmt = stmt.limit(self.MAX_ROWS)
-
-        with self.session_factory() as session, session.begin():
-            db_extensions = session.execute(stmt).scalars().all()
-
-            return [
-                mapper_registry.from_db_entity(db_extension)
-                for db_extension in db_extensions
-            ]
+        return func.int8range(
+            orm.Location.start,
+            orm.Location.end,
+            "[]",
+        ).op("&&")(func.int8range(start, stop, "[]"))
 
     def delete_extensions(
         self,
@@ -362,82 +128,3 @@ class PostgresObjectStore(Storage):
         with self.session_factory() as session, session.begin():
             result = session.execute(stmt)
             return result.rowcount or 0
-
-    def search_alleles(
-        self,
-        refget_accession: str,
-        start: int,
-        stop: int,
-        page_size: int = 1000,
-        cursor: str | None = None,
-    ) -> AlleleSearchPage:
-        """Find all Alleles that are located within the specified interval.
-
-        The interval is the closed range [start, stop] on the sequence identified by
-        the RefGet SequenceReference accession (`SQ.*`). Both `start` and `stop` are
-        inclusive and represent inter-residue positions.
-
-        Uses keyset pagination, meaning that altering the page size while looping through
-        successive cursors will effectively nullify the search loop.
-
-        Currently, any variation which overlaps the queried region is returned.
-
-        Raises an error if
-        * `start` or `end` are negative
-        * `end` > `start`
-
-        Todo (see Issue #338):
-        * define alternate match modes (partial/full overlap/contained/etc)
-        * define behavior for LSE indels and for alternative types of state (RLEs)
-
-        :param refget_accession: refget accession (e.g. `"SQ.IW78mgV5Cqf6M24hy52hPjyyo5tCCd86"`)
-        :param start: Inclusive, inter-residue start position of the interval
-        :param stop: Inclusive, inter-residue end position of the interval
-        :param page_size: Max # of results to return
-        :param cursor: Opaque key indicating start location for query in pagination
-        :return: Results page including variants and a cursor for next result page, if available
-        :raise InvalidSearchParamsError: if above search param requirements are violated
-        """
-        if start < 0 or stop < 0 or start > stop:
-            raise InvalidSearchParamsError
-        seek_start: int | None = None
-        seek_id: str | None = None
-        if cursor:
-            seek_start, seek_id = self._decode_search_cursor(cursor)
-
-        with self.session_factory() as session:
-            stmt = (
-                select(orm.Allele)
-                .options(
-                    joinedload(orm.Allele.location).joinedload(
-                        orm.Location.sequence_reference
-                    )
-                )
-                .join(orm.Location)
-                .join(orm.SequenceReference)
-                .where(
-                    orm.SequenceReference.id == refget_accession,
-                    func.int8range(orm.Location.start, orm.Location.end, "[]").op("&&")(
-                        func.int8range(start, stop, "[]")
-                    ),
-                )
-                .order_by(orm.Location.start, orm.Allele.id)
-                .limit(page_size)
-            )
-
-            # seek predicate -- assumes ORDER BY location.start ASC, allele.id ASC
-            if seek_start is not None and seek_id is not None:
-                stmt = stmt.where(
-                    or_(
-                        orm.Location.start > seek_start,
-                        and_(orm.Location.start == seek_start, orm.Allele.id > seek_id),
-                    )
-                )
-
-            page_db = session.scalars(stmt).all()
-            items = [mapper_registry.from_db_entity(a) for a in page_db]
-            if not page_db:
-                return AlleleSearchPage(items=[], next_cursor=None)
-            last = page_db[-1]
-            next_cursor = self._encode_search_cursor(last.location.start, last.id)
-            return AlleleSearchPage(items=items, next_cursor=next_cursor)
